+++ pytket/docs/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+Features:
+
+- Add `passes.CondCombinePass` which combines sequences of conditional gates into conditional boxes.
+
 ## 2.15.0 (March 2026)
 
 Fixes: